backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
python-dotenv
apscheduler
google-generativeai
PyPDF2
reportlab
pytest
PyJWT
